Finalizados servicos REST e conexao com DB

+++ src/main/java/com/hugoalves/sales/service/ClienteService.java
@@ -3,11 +3,13 @@ package com.hugoalves.sales.service;
 import com.hugoalves.sales.model.Cliente;
 import com.hugoalves.sales.model.repositories.ClienteRepository;
 import com.hugoalves.sales.service.exceptions.CnpjExistenteException;
+import com.hugoalves.sales.service.exceptions.CnpjInexistenteException;
+import com.hugoalves.sales.service.exceptions.ListaDeClientesVaziaException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
+
 
 @Service
 public class ClienteService {
@@ -23,15 +25,28 @@ public class ClienteService {
         }
     }
 
-    public List<Cliente> findAll(){
-        return repository.findAll();
+    public List<Cliente> findAll() throws ListaDeClientesVaziaException {
+        List<Cliente> clientes = repository.findAll();
+        if (clientes.isEmpty()){
+            throw new ListaDeClientesVaziaException();
+        } else {
+            return clientes;
+        }
     }
 
-    public Optional<Cliente> findByCnpj(String cnpj){
-        return repository.findByCnpj(cnpj);
+    public Cliente findByCnpj(String cnpj){
+        if (!repository.existsByCnpj(cnpj)){
+            throw new CnpjInexistenteException(cnpj);
+        } else {
+            return repository.findByCnpj(cnpj);
+        }
     }
 
     public void deleteByCnpj(String cnpj){
-        repository.deleteByCnpj(cnpj);
+        if (!repository.existsByCnpj(cnpj)){
+            throw new CnpjInexistenteException(cnpj);
+        } else {
+            repository.deleteByCnpj(cnpj);
+        }
     }
 }
